app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["talesmm14 <talesmelquiades@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.1"
django-environ = "^0.11.2"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
whitenoise = "^6.6.0"
drf-yasg = "^1.21.7"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
